fix: UI/UX audit fixes - hero, spacing, blog, services, contact metadata
- Reduce homepage hero padding and remove terminal animation delay
- Normalize section spacing (services, FAQ, founder) to py-16 lg:py-20
- Fix blog hero text visibility (add text-white, fix padding)
- Remove ScrollReveal from services Our Promise cards (fix mobile visibility)
- Extract ContactFormSection client component, keep contact page as Server Component with metadata
